// src/app/clearing/models/itemlicense.ts
import { GlobalVariables } from '../../core/models/globalvariables';

export class ItemLicense {

  lic_pkid: string;
  lic_job_id: string;
  lic_itm_id: string;
  lic_reg_no: string;
  lic_reg_date: string;
  lic_slno_parte: string;
  lic_exp_qty: number;
  lic_slno_partc: string;
  lic_ind_imported: string;
  lic_desc: string;
  lic_qty: number;
  lic_unit_id: string;
  lic_unit_code: string;
  lic_unit_name: string;
  lic_ctr: number;
  lic_deec: boolean;
   

  rec_mode: string;
  rec_category: string;
  _globalvariables: GlobalVariables;
}